quickly fixed tests to work with the new state
