+++ tests/components/Dashboard.test.js
@@ -1,15 +1,11 @@
 import React from 'react';
-import { Text } from 'react-native';
 
-import cloneDeep from 'lodash/cloneDeep';
 import configureMockStore from 'redux-mock-store';
 import { shallow } from 'enzyme';
 import { shallowToJson } from 'enzyme-to-json';
 
 import Dashboard from '../../src/components/Dashboard';
 import initialState from '../../src/constants/initialState';
-import lifecycles from '../../src/constants/lifecycles';
-import Loading from '../../src/components/Loading';
 
 const mockStore = configureMockStore();
 
@@ -18,28 +14,5 @@ describe('<Dashboard />', () => {
     const store = mockStore(initialState);
     const wrapper = shallow(<Dashboard store={store} />).shallow();
     expect(shallowToJson(wrapper)).toMatchSnapshot();
-  });
-  it('shows <Loading /> when status is loading', () => {
-    const modifiedState = cloneDeep(initialState);
-    modifiedState.appState.status = lifecycles.LOADING;
-
-    const store = mockStore(modifiedState);
-    const wrapper = shallow(<Dashboard store={store} />).shallow();
-
-    expect(wrapper.find(Loading)).toHaveLength(1);
-    expect(wrapper.find(Text)).toHaveLength(0);
-  });
-  it('does not show <Loading /> when status is not loading and shows right displayName', () => {
-    const modifiedState = cloneDeep(initialState);
-    modifiedState.appState.status = lifecycles.DATA_LOADED;
-    modifiedState.auth.user.displayName = 'bobby';
-
-    const store = mockStore(modifiedState);
-    const wrapper = shallow(<Dashboard store={store} />).shallow();
-
-    expect(wrapper.find(Loading)).toHaveLength(0);
-    expect(wrapper.find(Text)).toHaveLength(1);
-    // TODO: figure out a better way to ensure that text is properly set
-    expect(wrapper.find(Text).children().last().text()).toEqual('bobby');
   });
 });
